+++ app/Elasticsearch/Repositories/ItemRepository.php
@@ -449,6 +449,45 @@ class ItemRepository extends TranslatableRepository
         return $body;
     }
 
+    public function reindexAllLocales(): int
+    {
+        $processedCount = 0;
+        $locales = collect($this->locales);
+
+        $this
+            ->modelClass::with(['images', 'authorities:id', 'translations', 'tagged'])
+            ->chunk(1000, function($items) use (&$locales, &$processedCount) {
+                $processedCount += $items->count();
+
+                $locales->flatMap(function ($locale) use ($items) {
+                    return $items->flatMap(function ($item) use ($locale) {
+                        return [
+                            // Action
+                            [
+                                'index' => [
+                                    '_index' => $this->getLocalizedIndexName($locale),
+                                    '_id' => $item->getKey(),
+                                ]
+                            ],
+                    
+                            // Data
+                            $item->getIndexedData($locale)
+                        ];
+                    });
+                })
+                ->pipe(function ($bulkIndexBody) use (&$processedCount) {
+                    $this->elasticsearch->bulk(['body' => $bulkIndexBody->toArray()]);
+
+                    // Progress report
+                    if (app()->runningInConsole()) {
+                        echo date('h:i:s'). " " . $processedCount . "\n";
+                    }
+                });
+            });
+
+        return $processedCount;
+    }
+
     protected function getIndexConfig(string $locale  = null): array
     {
         return config('elasticsearch.index.items')[$this->getLocale($locale)];