+++ database/migrations/2021_01_14_134218_add_index_to_item_translations.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddIndexToItemTranslations extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('item_translations', function (Blueprint $table) {
+            $table->index(['item_id']);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('item_translations', function (Blueprint $table) {
+            $table->dropIndex(['item_id']);
+        });
+    }
+}