+++ public/ckeditor/plugins/slick/lang/en.js
@@ -1,0 +1,9 @@
+'use strict';
+
+(function (CKEDITOR) {
+    CKEDITOR.plugins.setLang('slick', 'en', {
+        title: 'Slick carousel',
+        imagesRequired: 'Please define at least one image.',
+        height: 'Height, use format: value+unit',
+    });
+})(CKEDITOR);