+++ resources/js/app.js
@@ -1,4 +1,5 @@
 require('./bootstrap')
+require( 'slick-carousel');
 
 const jQueryBridget = require('jquery-bridget')
 const Isotope = require('isotope-layout')
@@ -22,10 +23,32 @@ require('jquery.easing')
 require('./components/searchbar')
 require('./components/clipboard-button')
 
-const Vue = window.Vue = require('vue');
+//const Vue = window.Vue = require('vue');
 
 
 // color input for catalog-color filter
+
+// Vue components
+window.Vue = require('vue')
+Vue.component('user-collections-nav-link', require('./components/user-collections/NavLink.vue').default);
+Vue.component('user-collections-favourite-button', require('./components/user-collections/FavouriteButton.vue').default);
+Vue.component('user-collections-clear-button', require('./components/user-collections/ClearButton.vue').default);
+Vue.component('color-widget', require('./components/ColorWidget.vue').default);
+Vue.component('year-slider', require('./components/YearSlider.vue').default);
+
 Vue.component('slider', require('./components/vue/slider').default);
 Vue.component('color-slider', require('./components/vue/color-slider').default);
 
+
+/**
+ * Next, we will create a fresh Vue application instance and attach it to
+ * the page. Then, you may begin adding components to this application
+ * or customize the JavaScript scaffolding to fit your unique needs.
+ */
+
+ const app = new Vue({
+    el: '#app',
+    data: {
+        userCollectionsStore: require('./components/user-collections/store')
+    }
+});