+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Added
+- add es:migrate command
+### Changed
+- Use aliased indexes for es:setup
+- Speed up re-index for Items
+
+### Fixed
+- Do not load JS for "load-more" on single-page results
+- Fix zoom for related work in non-SK locale
+
+## [2.16.0] - 2021-01-15
+### Added
+- user collections
+### Changed
+- Do not specify default (develop) branch for test deploys
+- index is_for_reproductions in Elastic
+- exclude Cierne diery from reproductions
+
+### Changed
+- Make (reproduction) notices editable in admin
+
+## [2.15.1] - 2021-01-08
+### Changed
+- use debugbar in dev env
+- eager load articles
+- Bump ini from 1.3.5 to 1.3.7
 
 ## [2.15.0] - 2020-12-18
 ### Added
@@ -22,6 +48,7 @@ All notable changes to this project will be documented in this file[^1].
 - SGP into list of galleries in the info page
 - Run tests on GitHub Actions
 - autocomplete for item-author and item-authority-role fiels
+- ckeditor allows to add slick
 
 ## [2.13.1] - 2020-12-07
 ### Fixed